Hyperskill-Java-advanced/dividir-chocolate/src/main/java/com/futurelabs/RevertElements.java
package com.futurelabs;

import java.util.Arrays;
import java.util.Scanner;

public class RevertElements {
    // implement me
    public static int[] rotate(int[] arr, int steps) {
        for (int i = 0; i < steps; i++) {
            int lastElement = arr[arr.length - 1];
            for (int j = arr.length - 1; j > 0; j--) {
                arr[j] = arr[j - 1];
            }
            arr[0] = lastElement;
        }
        return arr;
    }

    // do not change code below
    public static void main(String[] args) {
        Scanner scanner = new Scanner(System.in);
        int[] arr = Arrays.stream(scanner.nextLine().split(" "))
                .mapToInt(Integer::parseInt)
                .toArray();

        int steps = Integer.parseInt(scanner.nextLine());

        rotate(arr, steps);

        for (int i : arr) {
            System.out.print(i + " ");
        }
    }
}
